Test non-classifier tests without scikit-learn installed (#78)
* Test non-classifier tests without scikit-learn installed

Also place all imports at top of the files, raising import errors when
needed

* Use coveragerc file to exclude debug lines from testing
