+++ .coveragerc
@@ -1,0 +1,5 @@
+[report]
+exclude_lines:
+    pragma: no cover
+    def __repr__
+    if verbose:
