+++ quantulum3/tests/test_setup.py
@@ -8,18 +8,18 @@
 import os
 import re
 import json
-import pickle
-import urllib.request
 import unittest
 
-# Dependences
-import wikipedia
+# Dependencies
+try:
+    import wikipedia
+except ImportError:
+    wikipedia = None
 
 # Quantulum
 from .. import load
 from .. import parser as p
 from .. import classes as cls
-from .. import classifier as clf
 
 COLOR1 = '\033[94m%s\033[0m'
 COLOR2 = '\033[91m%s\033[0m'
@@ -36,6 +36,9 @@ def wiki_test(page='CERN'):  # pragma: no cover
         Hubble_Space_Telescope,
         Herschel_Space_Observatory
     """
+    if not wikipedia:
+        print("Cannot activate wiki_test. Please install the package wikipedia first.")
+        return
 
     content = wikipedia.page(page).content
     parsed = p.parse(content)
@@ -141,7 +144,7 @@ def load_expand_tests():
 
 
 ################################################################################
-class EndToEndTests(unittest.TestCase):
+class SetupTest(unittest.TestCase):
     """Test suite for the quantulum3 project."""
 
     def test_load_tests(self):
@@ -149,47 +152,6 @@ class EndToEndTests(unittest.TestCase):
         self.assertFalse(load_quantity_tests(True) is None)
         self.assertFalse(load_quantity_tests(False) is None)
         self.assertFalse(load_expand_tests() is None)
-
-    def test_parse_classifier(self):
-        """ Test that parsing works with classifier usage """
-        all_tests = load_quantity_tests(False) + load_quantity_tests(True)
-        # forcedly activate classifier
-        clf.USE_CLF = True
-        for test in sorted(all_tests, key=lambda x: len(x['req'])):
-            quants = p.parse(test['req'])
-            self.assertEqual(
-                quants, test['res'],
-                "{} \n {}".format([quant.__dict__ for quant in quants],
-                                  [quant.__dict__ for quant in test['res']]))
-
-    def test_parse_no_classifier(self):
-        """ Test that parsing works without classifier usage """
-        all_tests = load_quantity_tests(False)
-        # forcedly deactivate classifier
-        clf.USE_CLF = False
-        for test in sorted(all_tests, key=lambda x: len(x['req'])):
-            quants = p.parse(test['req'])
-            self.assertEqual(
-                quants, test['res'], "\nExcpected: {1} \nGot: {0}".format(
-                    [quant.__dict__ for quant in quants],
-                    [quant.__dict__ for quant in test['res']]))
-
-    def test_training(self):
-        """ Test that classifier training works """
-        # TODO - update test to not overwirte existing clf.pickle and wiki.json files.
-        # Test that no errors are thrown during training
-        obj = clf.train_classifier(download=False, store=False)
-        # Test that the classifier works with the currently downloaded data
-        clf.TFIDF_MODEL, clf.CLF, clf.TARGET_NAMES = obj['tfidf_model'], obj[
-            'clf'], obj['target_names']
-        # Don't run with ambiguities because result is non-detemernistic
-        self.test_parse_no_classifier()
-
-    def test_expand(self):
-        all_tests = load_expand_tests()
-        for test in all_tests:
-            result = p.inline_parse_and_expand(test['req'])
-            self.assertEqual(result, test['res'])
 
     def test_build_script(self):
         """ Test that the build script has run correctly """
@@ -199,21 +161,6 @@ class EndToEndTests(unittest.TestCase):
             self.assertEqual(
                 load.COMMON_WORDS[length], word_set,
                 "Build script has not been run since change to critical files")
-
-    def test_classifier_up_to_date(self):
-        """ Test that the classifier has been built with the latest version of scikit-learn """
-        path = os.path.join(load.TOPDIR, 'clf.pickle')
-        with open(path, 'rb') as clf_file:
-            obj = pickle.load(clf_file, encoding='latin1')
-        clf_version = obj['scikit-learn_version']
-        with urllib.request.urlopen(
-                "https://pypi.org/pypi/scikit-learn/json") as response:
-            cur_version = json.loads(
-                response.read().decode('utf-8'))['info']['version']
-        self.assertEqual(
-            clf_version, cur_version,
-            "Classifier has been built with scikit-learn version {}, while the newest version is {}. Please update scikit-learn."
-            .format(clf_version, cur_version))
 
 
 ################################################################################
